Fixed bug in met_cen_lat_in and met_cen_lon_in user definitions.

+++ serial/src/mcipparm_mod.f90
@@ -284,7 +284,7 @@ MODULE mcipparm
 
   INTEGER :: dx_in, dy_in    ! user input  meteorological model resolution
 
-  INTEGER :: met_cen_lat_in, met_cen_lon_in    ! user input reference center latitude and longitude
+  REAL    :: met_cen_lat_in, met_cen_lon_in    ! user input reference center latitude and longitude
  
   INTEGER :: lpv             ! user input: 0 = Do not compute or output PV
                              !             1 = Compute and output PV
